+++ linopy/io.py
@@ -52,36 +52,18 @@ def objective_write_linear_terms(df, f, batch, batch_size):
     return batch
 
 
-def objective_write_cross_terms(quadratic, f, batch, batch_size):
+def objective_write_quad_terms(quadratic, f, batch, batch_size):
     """
     Write the cross terms of the objective to a file.
     """
-    is_cross = quadratic.vars1 != quadratic.vars2
-    cross = quadratic[is_cross]
-    coeffs = cross.coeffs.values
-    vars1 = cross.vars1.values
-    vars2 = cross.vars2.values
+    coeffs = quadratic.coeffs.values
+    vars1 = quadratic.vars1.values
+    vars2 = quadratic.vars2.values
     for idx in range(len(coeffs)):
         coeff = coeffs[idx]
         var1 = vars1[idx]
         var2 = vars2[idx]
         batch.append(f"{coeff:+.12g} x{var1} * x{var2}\n")
-        batch = handle_batch(batch, f, batch_size)
-    return batch
-
-
-def objective_write_quad_terms(quadratic, f, batch, batch_size):
-    """
-    Write the quadratic terms of the objective to a file.
-    """
-    is_cross = quadratic.vars1 != quadratic.vars2
-    quad = quadratic[~is_cross]
-    coeffs = quad.coeffs.values
-    vars = quad.vars1.values
-    for idx in range(len(coeffs)):
-        coeff = coeffs[idx]
-        var = vars[idx]
-        batch.append(f"{coeff:+.12g} x{var} ^ 2\n")
         batch = handle_batch(batch, f, batch_size)
     return batch
 
@@ -117,7 +99,6 @@ def objective_to_file(m, f, log=False, batch_size=10000):
             batch.append("+ [\n")
             quadratic = df[~is_linear]
             quadratic = quadratic.assign(coeffs=2 * quadratic.coeffs)
-            batch = objective_write_cross_terms(quadratic, f, batch, batch_size)
             batch = objective_write_quad_terms(quadratic, f, batch, batch_size)
             batch.append("] / 2\n")
 
@@ -249,7 +230,7 @@ def binaries_to_file(m, f, log=False, batch_size=1000):
         f.writelines(batch)
 
 
-def integers_to_file(m, f, log=False, batch_size=1000):
+def integers_to_file(m, f, log=False, batch_size=1000, integer_label="general"):
     """
     Write out integers of a model to a lp file.
     """
@@ -257,7 +238,7 @@ def integers_to_file(m, f, log=False, batch_size=1000):
     if not len(list(names)):
         return
 
-    f.write("\n\ninteger\n\n")
+    f.write(f"\n\n{integer_label}\n\n")
     if log:
         names = tqdm(
             list(names),
@@ -277,7 +258,7 @@ def integers_to_file(m, f, log=False, batch_size=1000):
         f.writelines(batch)
 
 
-def to_file(m, fn):
+def to_file(m, fn, integer_label="general"):
     """
     Write out a model to a lp or mps file.
     """
@@ -297,7 +278,13 @@ def to_file(m, fn):
             constraints_to_file(m, f, log=log, batch_size=batch_size)
             bounds_to_file(m, f, log=log, batch_size=batch_size)
             binaries_to_file(m, f, log=log, batch_size=batch_size)
-            integers_to_file(m, f, log=log, batch_size=batch_size)
+            integers_to_file(
+                m,
+                f,
+                log=log,
+                batch_size=batch_size,
+                integer_label=integer_label,
+            )
             f.write("end\n")
 
             logger.info(f" Writing time: {round(time.time()-start, 2)}s")
@@ -570,8 +557,18 @@ def to_netcdf(m, *args, **kwargs):
     """
 
     def with_prefix(ds, prefix):
-        ds = ds.rename({d: f"{prefix}-{d}" for d in [*ds.dims, *ds]})
+        to_rename = set([*ds.dims, *ds.coords, *ds])
+        ds = ds.rename({d: f"{prefix}-{d}" for d in to_rename})
         ds.attrs = {f"{prefix}-{k}": v for k, v in ds.attrs.items()}
+
+        # Flatten multiindexes
+        for dim in ds.dims:
+            if isinstance(ds[dim].to_index(), pd.MultiIndex):
+                prefix_len = len(prefix) + 1  # leave original index level name
+                names = [n[prefix_len:] for n in ds[dim].to_index().names]
+                ds = ds.reset_index(dim)
+                ds.attrs[f"{dim}_multiindex"] = list(names)
+
         return ds
 
     vars = [
@@ -581,11 +578,15 @@ def to_netcdf(m, *args, **kwargs):
         with_prefix(con.data, f"constraints-{name}")
         for name, con in m.constraints.items()
     ]
-    obj = [with_prefix(m.objective.data, "objective")]
+    objective = m.objective.data
+    objective = objective.assign_attrs(sense=m.objective.sense)
+    if m.objective.value is not None:
+        objective = objective.assign_attrs(value=m.objective.value)
+    obj = [with_prefix(objective, "objective")]
     params = [with_prefix(m.parameters, "params")]
 
     scalars = {k: getattr(m, k) for k in m.scalar_attrs}
-    ds = xr.merge(vars + cons + obj + params)
+    ds = xr.merge(vars + cons + obj + params, combine_attrs="drop_conflicts")
     ds = ds.assign_attrs(scalars)
     ds.attrs = non_bool_dict(ds.attrs)
 
@@ -622,30 +623,57 @@ def read_netcdf(path, **kwargs):
     m = Model()
     ds = xr.load_dataset(path, **kwargs)
 
+    def has_prefix(k, prefix):
+        return k.rsplit("-", 1)[0] == prefix
+
+    def remove_prefix(k, prefix):
+        return k[len(prefix) + 1 :]
+
     def get_prefix(ds, prefix):
-        ds = ds[[k for k in ds if k.startswith(prefix)]]
-        ds = ds.rename({d: d.split(prefix + "-", 1)[1] for d in [*ds.dims, *ds]})
+        ds = ds[[k for k in ds if has_prefix(k, prefix)]]
+        multiindexes = []
+        for dim in ds.dims:
+            for name in ds.attrs.get(f"{dim}_multiindex", []):
+                multiindexes.append(prefix + "-" + name)
+        ds = ds.drop_vars(set(ds.coords) - set(ds.dims) - set(multiindexes))
+        to_rename = set([*ds.dims, *ds.coords, *ds])
+        ds = ds.rename({d: remove_prefix(d, prefix) for d in to_rename})
         ds.attrs = {
-            k.split(prefix + "-", 1)[1]: v
+            remove_prefix(k, prefix): v
             for k, v in ds.attrs.items()
-            if k.startswith(prefix + "-")
+            if has_prefix(k, prefix)
         }
+
+        for dim in ds.dims:
+            if f"{dim}_multiindex" in ds.attrs:
+                names = ds.attrs.pop(f"{dim}_multiindex")
+                ds = ds.set_index({dim: names})
+
         return ds
 
-    vars = get_prefix(ds, "variables")
-    var_names = list({k.split("-", 1)[0] for k in vars})
-    variables = {k: Variable(get_prefix(vars, k), m, k) for k in var_names}
+    vars = [k for k in ds if k.startswith("variables")]
+    var_names = list({k.rsplit("-", 1)[0] for k in vars})
+    variables = {}
+    for k in sorted(var_names):
+        name = remove_prefix(k, "variables")
+        variables[name] = Variable(get_prefix(ds, k), m, name)
+
     m._variables = Variables(variables, m)
 
-    cons = get_prefix(ds, "constraints")
-    con_names = list({k.split("-", 1)[0] for k in cons})
-    constraints = {k: Constraint(get_prefix(cons, k), m, k) for k in con_names}
+    cons = [k for k in ds if k.startswith("constraints")]
+    con_names = list({k.rsplit("-", 1)[0] for k in cons})
+    constraints = {}
+    for k in sorted(con_names):
+        name = remove_prefix(k, "constraints")
+        constraints[name] = Constraint(get_prefix(ds, k), m, name)
     m._constraints = Constraints(constraints, m)
 
     objective = get_prefix(ds, "objective")
-    m._objective = LinearExpression(objective, m)
+    m.objective = LinearExpression(objective, m)
+    m.objective.sense = objective.attrs.pop("sense")
+    m.objective._value = objective.attrs.pop("value", None)
 
-    m._parameters = get_prefix(ds, "parameter")
+    m.parameters = get_prefix(ds, "parameter")
 
     for k in m.scalar_attrs:
         setattr(m, k, ds.attrs.get(k))